backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
langchain
langchain-google-genai
python-dotenv
chromadb
pydantic
pypdf
